Moved files around a bit, and added a low level optimization
